feat: (ui) reuseable publish button
