+++ aether-ui/aether/ui/assets/apps/pipeline/PipelineList.jsx
@@ -4,7 +4,7 @@ import { FormattedMessage } from 'react-intl'
 import moment from 'moment'
 
 import { PROJECT_NAME } from '../utils/constants'
-import { NavBar } from '../components'
+import { NavBar, PublishButton } from '../components'
 
 import NewPipeline from './NewPipeline'
 import { addPipeline, selectedPipelineChanged, getPipelines, fetchPipelines } from './redux'
@@ -25,7 +25,6 @@ class PipelineList extends Component {
   }
 
   componentWillReceiveProps (nextProps) {
-    console.log(nextProps.pipelineList)
     if (nextProps.error !== this.props.error) {
       // TODO: handle errors
     }
@@ -61,43 +60,40 @@ class PipelineList extends Component {
     return this.props.pipelineList.map(pipeline => (
       <div
         key={pipeline.id}
-        className='pipeline-preview'
-        onClick={() => { this.onSelectPipeline(pipeline) }}>
-        <h2 className='preview-heading'>{pipeline.name}</h2>
+        className='pipeline-preview'>
+        <div
+          onClick={() => { this.onSelectPipeline(pipeline) }}>
+          <h2 className='preview-heading'>{pipeline.name}</h2>
 
-        <div className='summary-entity-types'>
-          <span className='badge badge-b badge-big'>
-            { pipeline.entity_types ? pipeline.entity_types.length : 0 }
-          </span>
-          <FormattedMessage
-            id='pipeline.list.entity.types'
-            defaultMessage='Entity-Types'
-          />
-        </div>
+          <div className='summary-entity-types'>
+            <span className='badge badge-b badge-big'>
+              { pipeline.entity_types ? pipeline.entity_types.length : 0 }
+            </span>
+            <FormattedMessage
+              id='pipeline.list.entity.types'
+              defaultMessage='Entity-Types'
+            />
+          </div>
 
-        <div className='summary-errors'>
-          <span className='badge badge-b badge-big'>
-            { pipeline.mapping_errors ? pipeline.mapping_errors.length : 0 }
-          </span>
-          <FormattedMessage
-            id='pipeline.list.errors'
-            defaultMessage='Errors'
-          />
+          <div className='summary-errors'>
+            <span className='badge badge-b badge-big'>
+              { pipeline.mapping_errors ? pipeline.mapping_errors.length : 0 }
+            </span>
+            <FormattedMessage
+              id='pipeline.list.errors'
+              defaultMessage='Errors'
+            />
+          </div>
         </div>
         <div className='pipeline-publish'>
           <div className='status-publish'>
             <FormattedMessage
               id='pipeline.list.publish-status'
-              defaultMessage={pipeline.published_on ? `published on ${moment(new Date(pipeline.published_on)).format('MMMM DD')}` :
-              'not published'}
+              defaultMessage={pipeline.published_on ? `Published on ${moment(new Date(pipeline.published_on)).format('MMMM DD')}`
+                : 'Not published'}
             />
           </div>
-          <button type='button' className='btn btn-w btn-publish'>
-            <FormattedMessage
-              id='pipeline.navbar.breadcrumb.publish'
-              defaultMessage='Publish pipeline'
-            />
-          </button>
+          <PublishButton pipeline={pipeline} className='btn btn-w btn-publish' />
         </div>
 
       </div>
