+++ aether-ui/aether/ui/assets/css/base/_nav-bar.scss
@@ -3,7 +3,7 @@
   transition: padding .2s;
   justify-content: flex-start;
   align-items: baseline;
-  z-index: 1;
+  z-index: 3;
 }
 
 .top-nav:hover + .pipeline {
